refactor: Implement new folder structure for activities, services, models, and utils
