+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
         android:theme="@style/Base.Theme.App"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".activities.MainActivity"
             android:exported="true"
             android:theme="@style/Base.Theme.Main">
             <intent-filter>
@@ -27,22 +27,22 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".RegistroUser"
+            android:name=".activities.RegistroUser"
             android:exported="false"
             android:label="@string/act_title_Registro"
             android:theme="@style/AppTheme"
-            android:parentActivityName=".MainActivity"/>
+            android:parentActivityName=".activities.MainActivity"/>
         <activity
-            android:name=".ListadoNotas"
+            android:name=".activities.ListadoNotas"
             android:exported="false"
             android:theme="@style/AppTheme"
             android:label="@string/act_title_ListadoNotas"/>
         <activity
-            android:name=".NuevaNota"
+            android:name=".activities.NuevaNota"
             android:exported="false"
             android:label="Crea una Nota"
             android:theme="@style/AppTheme"
-            android:parentActivityName=".ListadoNotas" />
+            android:parentActivityName=".activities.ListadoNotas" />
     </application>
 
 </manifest>
