+++ app/src/main/res/layout/activity_listado_notas.xml
@@ -5,7 +5,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ListadoNotas"
+    tools:context=".activities.ListadoNotas"
     android:background="@drawable/background">
 
   <Button
