+++ app/src/main/res/layout/activity_nueva_nota.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".NuevaNota"
+    tools:context=".activities.NuevaNota"
     android:background="@drawable/background">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
