+++ app/src/main/res/layout/activity_registro_user.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RegistroUser"
+    tools:context=".activities.RegistroUser"
     android:background="@drawable/background">
 
     <TextView
